Rename module to 'legacy' as per LIP-0050
