+++ src/application/modules/index.ts
@@ -12,4 +12,4 @@
  * Removal or modification of this copyright notice is prohibited.
  */
 
-export * from './legacy_account';
+export * from './legacy';
